requirements.txt
amqp==5.3.1
annotated-types==0.7.0
anyio==4.8.0
aplus==0.11.0
asdf==4.1.0
asdf-astropy==0.7.0
asdf_coordinates_schemas==0.3.0
asdf_standard==1.1.1
asdf_transform_schemas==0.5.0
asdf_wcs_schemas==0.4.0
asgiref==3.8.1
astropy==6.1.7
astropy-iers-data==0.2025.2.3.0.32.42
async-timeout==5.0.1
attrs==25.1.0
backports.tarfile==1.2.0
billiard==4.2.1
blake3==1.0.4
celery==5.4.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cloudpickle==3.1.1
colorama==0.4.6
contourpy==1.3.1
cycler==0.12.1
dask==2025.1.0
defusedxml==0.7.1
Django==5.1.6
django-rest-framework==0.1.0
django-restframework==0.0.1
djangorestframework==3.15.2
exceptiongroup==1.2.2
filelock==3.17.0
fonttools==4.56.0
frozendict==2.4.6
fsspec==2025.2.0
future==1.0.0
gwcs==0.24.0
h11==0.14.0
httpcore==1.0.7
httpie==3.2.4
idna==3.10
importlib_metadata==8.6.1
jaraco.classes==3.4.0
jaraco.context==6.0.1
jaraco.functools==4.1.0
jmespath==1.0.1
kiwisolver==1.4.8
kombu==5.4.2
locket==1.0.0
markdown-it-py==3.0.0
matplotlib==3.10.0
mdurl==0.1.2
more-itertools==10.6.0
msgpack==1.1.0
multidict==6.1.0
narwhals==1.25.2
ndcube==2.3.0
ndindex==1.9.2
nest-asyncio==1.6.0
numexpr==2.10.2
numpy==2.1.3
packaging==24.2
pandas==2.2.3
partd==1.4.2
pillow==11.1.0
progressbar2==4.5.0
prompt_toolkit==3.0.50
psycopg2==2.9.10
py-cpuinfo==9.0.0
pyarrow==19.0.0
pydantic==2.10.6
pydantic_core==2.27.2
pyerfa==2.0.1.5
Pygments==2.19.1
pyparsing==3.2.1
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-utils==3.9.1
pytz==2025.1
pyvo==1.6
pywin32-ctypes==0.2.3
PyYAML==6.0.2
redis==5.2.1
requests==2.32.3
requests-toolbelt==1.0.0
rich==13.9.4
ruamel.yaml==0.18.10
ruamel.yaml.clib==0.2.12
scipy==1.15.1
semantic-version==2.10.0
six==1.17.0
sniffio==1.3.1
soupsieve==2.6
sqlparse==0.5.3
tabulate==0.9.0
toolz==1.0.0
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2025.1
uncertainties==3.2.2
urllib3==2.2.0
vaex-core==4.17.1
vine==5.1.0
wcwidth==0.2.13
webencodings==0.5.1
zipp==3.21.0
